+++ AutoGame1989/Vehicle.h
@@ -516,7 +516,7 @@ private:
     bool m_isFuelOn = true;
 
     float m_testVelocity = 0.0;
-    float m_testTimer = 0.0;
+    double m_testTimer = 0.0;
     double m_testTimerTotal = 0.0;
 
     float m_testEnginePower = 0.0;
